+++ src/core/tree_type.ml
@@ -6,6 +6,17 @@ open Common
 (** {3 Atomic pattern constructor representation} *)
 
 type pvar = int * int array
+type psym = Path.t * string
+type c_term =
+  | C_App of psym * c_term list  (* application *)
+  | C_Patt of pvar               (* pattern variable *)
+  | C_Sym of psym                (* global symbol *)
+
+let rec c_term : c_term pp = fun ppf t ->
+  match t with
+  | C_App(s,al) -> out ppf "%s [%a]" (snd s) (List.pp c_term ",") al
+  | C_Patt (i,_) -> out ppf "?%d" i
+  | C_Sym s -> out ppf "%s" (snd s)
 
 (** Representation of an atomic pattern constructor. *)
 module TC =
@@ -56,6 +67,10 @@ type tree_cond =
   | CondFV of pvar
   (** Are the (indexed) bound variables (which are free at the time of the
       checking) of the term at the given index in the array? *)
+  | CondEQ of (psym * pvar list) * (psym * pvar list)
+  (** user rule condition: are the two terms equivalent *)
+  | CondCHK of psym * c_term list
+  (** user rule predefined condition *)
 
 (** {b NOTE} that when performing a [tree_cond.CondFV] check, we
     are concerned about variables that were bound in the term being reduced
@@ -69,6 +84,11 @@ let tree_cond : tree_cond pp = fun ppf tc ->
       out ppf "Nl(%d[%a],%d[%a])"
         i (Array.pp int ",") xs j (Array.pp int ",") ys
   | CondFV(i, _) -> out ppf "Fv(%d)" i
+  | CondEQ((s1,a1), (s2,a2)) ->
+      out ppf "Eq(%s(%a),%s(%a))"
+        (snd s1) (List.pp int ",") (List.map fst a1)
+        (snd s2) (List.pp int ",") (List.map fst a2)
+  | CondCHK(op,al) -> out ppf "CHK %s[%a]" (snd op) (List.pp c_term " ") al
 
 (** Substitution of variables in a RHS. During the filtering process, some
     subterms of the filtered term may be stored in an array. Let [v] be that