+++ src/core/unif.ml
@@ -63,7 +63,9 @@ let rec type_app : ctxt -> term -> term list -> term option = fun c a ts ->
   | _, [] -> Some a
   | Prod(_,b), t::ts -> type_app c (subst b t) ts
   | LLet(_,d,b), t::ts ->
-      assert (Eval.pure_eq_modulo c d t); type_app c (subst b d) ts
+let _ = if not (Eval.pure_eq_modulo c d t) then
+log "LET ERROR??? %a -- %a" term a term t in
+      (*assert (Eval.pure_eq_modulo c d t);*) type_app c (subst b d) ts
   | _ ->
       match Eval.whnf c a, ts with
       | Prod(_,b), t::ts -> type_app c (subst b t) ts