+++ src/parsing/lpLexer.ml
@@ -100,6 +100,7 @@ type token =
   | TYPE_TERM
   | UNIF_RULE
   | VERBOSE
+  | WHEN
   | WHY3
   | WITH
 
@@ -305,6 +306,7 @@ let rec token ~allow_rocq_syntax lb =
   | "TYPE" -> TYPE_TERM
   | "unif_rule" -> UNIF_RULE
   | "verbose" -> VERBOSE
+  | "when" -> WHEN
   | "why3" -> WHY3
   | "with" -> WITH