+++ src/parsing/scope.ml
@@ -554,11 +554,26 @@ let patt_vars : p_term -> (string * int) list * string list =
 let scope_rule :
   ?find_sym:find_sym -> bool -> sig_state -> p_rule -> sym_rule =
   fun ?(find_sym=Sig_state.find_sym) ur ss
-      { elt = (p_lhs, p_rhs); pos = rule_pos} ->
+      { elt = (p_lhs, p_rhs, p_when); pos = rule_pos} ->
   (* Compute the set of pattern variables on both sides. *)
   let (pvs_lhs, nl) = patt_vars p_lhs in
   (* NOTE to reject non-left-linear rules check [nl = []] here. *)
   let (pvs_rhs, _ ) = patt_vars p_rhs in
+  (* input/output variables in condition *)
+  let pvs_when =
+    match p_when with
+    | P_None -> [],[]
+    | P_EQ (t1,t2) ->
+        let (p1,_) = patt_vars t1 and (p2,_) = patt_vars t2 in
+        (p1 @ p2), []
+    | P_CHK t -> 
+        match p_get_args t with
+          (* find_doubles op! n! t! d? before? middle? after? *)
+        | {elt=P_Iden({elt=(_,"#find_doubles");_},_);_}, [ty;op;u;t;d;b;m;a] ->
+            (List.concat_map (fun t -> fst (patt_vars t)) [ty;u;t]),
+            (List.concat_map (fun t -> fst (patt_vars t)) [d;b;m;a])
+        | _ -> fatal t.pos "Undefined constraint."
+  in
   (* Check that pattern variables of RHS that are in the LHS have the right
      arity. *)
   let check_arity (m,i) =
@@ -568,6 +583,8 @@ let scope_rule :
     with Not_found -> ()
   in
   List.iter check_arity pvs_rhs;
+  List.iter check_arity (fst pvs_when);
+  List.iter check_arity (snd pvs_when);
   (* [get_root t] returns the symbol at the root of the p_term [t]. *)
   let rec get_root t = get_root_after_pratt (Pratt.parse ~find_sym ss [] t)
   and get_root_after_pratt t =
@@ -599,9 +616,13 @@ let scope_rule :
            ; m_lhs_arities = Hashtbl.create 7
            ; m_lhs_names   = Hashtbl.create 7
            ; m_lhs_size    = 0
-           ; m_lhs_in_env  = nl @ List.map fst pvs_rhs }
+           ; m_lhs_in_env  =
+               nl @ List.map fst pvs_rhs @ List.map fst (snd pvs_when)}
     in
     let lhs = scope ~find_sym 0 mode ss Env.empty p_lhs in
+    let _ = match p_when with
+        P_CHK t -> ignore (scope ~find_sym 0 mode ss Env.empty t)
+      | _ -> () in
     match mode with
     | M_LHS{ m_lhs_indices; m_lhs_names; m_lhs_size; m_lhs_arities; _} ->
       let lhs = snd (get_args lhs) in
@@ -630,7 +651,40 @@ let scope_rule :
   let arity = List.length lhs in
   let f i = try Hashtbl.find names i with Not_found -> string_of_int i in
   let names = Array.init vars_nb f in
-  let r = {lhs; names; rhs; arity; arities; vars_nb; xvars_nb; rule_pos} in
+  let asPatt t = match unfold t with
+    | Patt (Some i, _, [||]) -> i
+    | _ -> fatal rule_pos "arguments should be patterns in condition [%a]."
+             term t in
+  let rec asRTerm: Term.term -> Term.r_term = fun t ->
+    match get_args t with
+    | Patt (Some i, _, [||]), [] -> R_Patt i
+    | Patt (None, _, _), _ ->
+        fatal rule_pos "anonymous pattern not allowed in constraint [%a]."
+          term t
+    | Patt _, _ ->
+        fatal rule_pos "pattern applicationn not allowed in constraint [%a]."
+          term t
+    | Symb s, tl -> R_App (s, List.map asRTerm tl)
+    | f,_ -> fatal rule_pos "Not allowed in constraint [%a]."
+               term f in
+  let asRTerm : Term.term -> sym * Term.r_term list = fun t ->
+    match asRTerm t with
+    | R_App (s,al) -> s,al
+    | _ -> fatal rule_pos "constraint should be a function call [%a]."
+             term t in
+  let p_asRTerm t = asRTerm (scope ~find_sym 0 mode ss Env.empty t) in
+  let pw2rw t =
+    let t = scope ~find_sym 0 mode ss Env.empty t in
+    match get_args t with
+      Symb s, args ->  (s, List.map asPatt args)
+    | _ -> fatal rule_pos "function should be a global symbol in [%a]." term t
+  in
+  let r_when = match p_when with
+    | P_None -> R_None
+    | P_EQ (t1,t2) -> R_EQ (pw2rw t1, pw2rw t2)
+    | P_CHK t -> let (f,al) = p_asRTerm t in R_CHK (f,al) in
+  let r = {lhs; names; rhs; arity; arities; vars_nb; xvars_nb;
+           rule_pos; r_when} in
   (sym,r)
 
 (** [scope_pattern ss env t] turns a parser-level term [t] into an actual term