+++ tests/export_raw_dk.sh
@@ -58,6 +58,8 @@ do
         require_nondkmident);;
         # module alias
         alias);;
+        # conditional rules
+        cond_rules*);;
         # proofs
         why3*|tutorial|try|tautologies|rewrite*|remove|natproofs|have|generalize|foo|comment_in_qid|apply|anonymous|admit|change|assumption|focus|assume|first_hyp|all_hyps);;
         # "open"